update 01/24/22 again

+++ StudentRecApi/src/main/java/com/student/util/CSVHelper.java
@@ -118,9 +118,9 @@ public class CSVHelper {
 	}
 	
 	public static boolean isNumeric(String number) {
-		int intValue = 0;
+		Long intValue = (long) 0;
 		try {
-			intValue = Integer.parseInt(number);
+			intValue = Long.parseLong(number);
 			return true;
 		} catch (NumberFormatException e) {
 			return false;
